MAN-636 test coverage wip
