+++ integration_tests/e2e/appointments/check-your-answers.cy.ts
@@ -23,6 +23,8 @@ import {
   completeAttendancePage,
   completeNotePage,
   checkPopHeader,
+  checkAppointmentSummary,
+  checkAppointmentUpdate,
 } from './imports'
 import { statusErrors } from '../../../server/properties'
 
@@ -156,108 +158,11 @@ describe('Check your answers then confirm the appointment', () => {
 
   describe('Change appointment values', () => {
     let cyaPage: AppointmentCheckYourAnswersPage
-    let typePage: AppointmentTypePage
-    let sentencePage: AppointmentSentencePage
-    let dateTimePage: AppointmentDateTimePage
-    let locationPage: AppointmentLocationPage
-    let repeatingPage: AppointmentRepeatingPage
-    let locationNotInListPage: AppointmentLocationNotInListPage
-    let notePage: AppointmentNotePage
     beforeEach(() => {
       loadPage({})
       cyaPage = new AppointmentCheckYourAnswersPage()
     })
-    it('should update the type when value is changed', () => {
-      cyaPage.getSummaryListRow(1).find('.govuk-link').click()
-      typePage = new AppointmentTypePage()
-      typePage.getRadio('type', 2).click()
-      typePage.getSubmitBtn().click()
-      cyaPage = new AppointmentCheckYourAnswersPage()
-      cyaPage.checkOnPage()
-      cyaPage.getSummaryListRow(1).find('.govuk-summary-list__value').should('contain.text', 'Home Visit to Case (NS)')
-    })
-    it('should update the sentence when value is changed', () => {
-      cyaPage.getSummaryListRow(2).find('.govuk-link').click()
-      sentencePage = new AppointmentSentencePage()
-      sentencePage.getElement(`#appointments-${crn}-${uuid}-eventId-2`).click()
-      sentencePage.getElement(`#appointments-${crn}-${uuid}-requirementId`).click()
-      sentencePage.getSubmitBtn().click()
-      cyaPage = new AppointmentCheckYourAnswersPage()
-      cyaPage.checkOnPage()
-      cyaPage
-        .getSummaryListRow(2)
-        .find('.govuk-summary-list__value')
-        .should('contain.text', 'ORA Community Order')
-        .should('contain.text', '12 days RAR, 1 completed')
-    })
-    it('should update the location when value is changed', () => {
-      cyaPage.getSummaryListRow(4).find('.govuk-link').click()
-      locationPage = new AppointmentLocationPage()
-      locationPage.getRadio('locationCode', 2).click()
-      locationPage.getSubmitBtn().click()
-      cyaPage.getSummaryListRow(4).find('.govuk-summary-list__value').should('contain.text', '102 Petty France')
-    })
-    it('should update the location by clicking on location not in list, then back link, then updating the value', () => {
-      cyaPage.getSummaryListRow(4).find('.govuk-link').click()
-      locationPage = new AppointmentLocationPage()
-      locationPage.getRadio('locationCode', 4).click()
-      locationPage.getSubmitBtn().click()
-      locationNotInListPage = new AppointmentLocationNotInListPage()
-      locationNotInListPage.getBackLink().click()
-      locationPage.getRadio('locationCode', 2).click()
-      locationPage.getSubmitBtn().click()
-      cyaPage.getSummaryListRow(4).find('.govuk-summary-list__value').should('contain.text', '102 Petty France')
-    })
-    it('should update the date/time when value is changed', () => {
-      const changedStart = '9:30am'
-      const changedEnd = '10:30am'
-      cyaPage.getSummaryListRow(5).find('.govuk-link').click()
-      dateTimePage = new AppointmentDateTimePage()
-      dateTimePage.getElement(`#appointments-${crn}-${uuid}-start`).select(changedStart)
-      dateTimePage.getElement(`#appointments-${crn}-${uuid}-end`).focus().select(changedEnd).tab()
-      dateTimePage.getSubmitBtn().click()
-      // Ignore warnings
-      dateTimePage.getSubmitBtn().click()
-      cyaPage
-        .getSummaryListRow(5)
-        .find('.govuk-summary-list__value li:nth-child(1)')
-        .invoke('text')
-        .then(text => {
-          const normalizedText = text.replace(/\s+/g, ' ').trim()
-          expect(normalizedText).to.include(`${dateWithYear(date)} from ${changedStart} to ${changedEnd}`)
-        })
-    })
-    it('should update the repeating appointment when value is changed', () => {
-      cyaPage.getSummaryListRow(6).find('.govuk-link').click()
-      repeatingPage = new AppointmentRepeatingPage()
-      repeatingPage.getElement(`#appointments-${crn}-${uuid}-repeating-2`).click()
-      repeatingPage.getSubmitBtn().click()
-      cyaPage.getSummaryListRow(6).find('.govuk-summary-list__value').should('contain.text', 'No')
-      cyaPage.getSummaryListRow(5).find('.govuk-summary-list__value li').should('have.length', 1)
-      cyaPage
-        .getSummaryListRow(5)
-        .find('.govuk-summary-list__value li:nth-child(1)')
-        .invoke('text')
-        .then(text => {
-          const normalizedText = text.replace(/\s+/g, ' ').trim()
-          expect(normalizedText).to.include(`${dateWithYear(date)} from ${startTime} to ${endTime}`)
-        })
-    })
-    it('should update the notes when value is changed', () => {
-      cyaPage.getSummaryListRow(7).find('.govuk-link').click()
-      const updatedNotes = 'Some updated notes'
-      notePage = new AppointmentNotePage()
-      notePage.getElement(`#notes`).focus().type(updatedNotes)
-      notePage.getSubmitBtn().click()
-      cyaPage.getSummaryListRow(7).find('.govuk-summary-list__value').should('contain.text', updatedNotes)
-    })
-    it('should update the sensitivity when value is changed', () => {
-      cyaPage.getSummaryListRow(8).find('.govuk-link').click()
-      notePage = new AppointmentNotePage()
-      notePage.getElement(`#appointments-${crn}-${uuid}-sensitivity-2`).click()
-      notePage.getSubmitBtn().click()
-      cyaPage.getSummaryListRow(8).find('.govuk-summary-list__value').should('contain.text', 'No')
-    })
+    checkAppointmentUpdate(cyaPage)
   })
   describe('Confirm this appointment', () => {
     let cyaPage: AppointmentCheckYourAnswersPage
